backend-api/requirements.txt
# Django and Core Framework
Django>=5.1.0
django-environ>=0.11.2

# Database and GeoDjango
psycopg2-binary>=2.9.7
django-extensions>=3.2.3

# GeoDjango dependencies  
GDAL>=3.6.4
Pillow>=10.0.0

# Django REST Framework
djangorestframework>=3.14.0
django-rest-framework-gis>=1.0
django-filter>=23.3
drf-spectacular>=0.27.0

# CORS handling for frontend
django-cors-headers>=4.3.0

# Async HTTP client for NLP services
aiohttp>=3.8.5
asyncio-throttle>=1.0.2

# Celery for async task processing
celery>=5.3.1
redis>=4.6.0

# Monitoring and logging
django-debug-toolbar>=4.2.0
structlog>=23.1.0

# Security and authentication
django-oauth-toolkit>=1.7.0
djangorestframework-jwt>=1.11.0
cryptography>=41.0.4

# Data processing and utilities
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0

# Development and testing
pytest>=7.4.2
pytest-django>=4.5.2
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
factory-boy>=3.3.0

# Code quality
flake8>=6.0.0
black>=23.7.0
isort>=5.12.0

# Production deployment
gunicorn>=21.2.0
whitenoise>=6.5.0

# Environment management
python-decouple>=3.8
django-storages>=1.14.0

# Geospatial libraries
Shapely>=2.0.1
Fiona>=1.9.4
